safari-600.7-branch/LayoutTests/fast/url/invalid-urls-utf8-expected.txt

This ensure we do not mess up the bytes when an URL is partially parsed and is invalid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL src should be http:///. Was http:/.
FAIL src should be https:///. Was https:/.
FAIL src should be ftp:///. Was ftp://.
FAIL src should be gopher:///. Was gopher://.
FAIL src should be ws:///. Was ws://.
FAIL src should be wss:///. Was wss://.
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS src is expected
PASS successfullyParsed is true

TEST COMPLETE
